old-api/requirements.txt
Flask==0.10.1
boto3==1.4.1
simplejson==3.10.0
logging==0.4.9.6
uuid==1.30
pyjwt==1.4.2
python-dotenv==0.6.1
pymongo==3.3.1
passlib==1.7.1
flasgger==0.8.0
pypubsub==3.3.0
